+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  * Send IP obtained through `javax.servlet.ServletRequest#getRemoteAddr()` in `context.request.socket.remote_address` 
    instead of parsing from headers (#889)
  * Added `ElasticApmAttacher.attach(String propertiesLocation)` to specify a custom properties location
+ * Logs message when `transaction_max_spans` has been exceeded (#849)
 
 ## Bug Fixes
  * JMS creates polling transactions even when the API invocations return without a message